backend/requirements.txt
flask
flask_sqlalchemy
pymysql
flask-cors
werkzeug
pyjwt
flask-socketio
flask-jwt-extended

